Test Unitaires, Integration et e2e

+++ back/src/main/java/com/openclassrooms/starterjwt/services/SessionService.java
@@ -27,6 +27,10 @@ public class SessionService {
     }
 
     public void delete(Long id) {
+        Session session = this.getById(id);
+        if (session == null) {
+            throw new NotFoundException();
+        }
         this.sessionRepository.deleteById(id);
     }
 
@@ -35,7 +39,11 @@ public class SessionService {
     }
 
     public Session getById(Long id) {
-        return this.sessionRepository.findById(id).orElse(null);
+        Session session = this.sessionRepository.findById(id).orElse(null);
+        if (session == null) {
+            throw new NotFoundException();
+        }
+        return session;
     }
 
     public Session update(Long id, Session session) {
@@ -50,7 +58,7 @@ public class SessionService {
             throw new NotFoundException();
         }
 
-        boolean alreadyParticipate = session.getUsers().stream().anyMatch(o -> o.getId().equals(userId));
+        boolean alreadyParticipate = session.getUsers().stream().anyMatch(u -> u.getId().equals(userId));
         if (alreadyParticipate) {
             throw new BadRequestException();
         }
@@ -66,7 +74,7 @@ public class SessionService {
             throw new NotFoundException();
         }
 
-        boolean alreadyParticipate = session.getUsers().stream().anyMatch(o -> o.getId().equals(userId));
+        boolean alreadyParticipate = session.getUsers().stream().anyMatch(user -> user.getId().equals(userId));
         if (!alreadyParticipate) {
             throw new BadRequestException();
         }
